Refactor Ledger class - create BlockImport for all actions connected with importing new block - create BlockExecution for all actions connected with block transactions - create BranchResolution for all actions connected with resolving branch - adjust test class to new classes
